LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/values/shape-outside-circle-004-expected.txt


FAIL test unit (inline): cm - circle(at 50cm) assert_equals: expected "circle(at 50cm 50%)" but got "circle(at 50cm center)"
PASS test unit (inline): cm - circle(at 50% 50cm)
PASS test unit (inline): cm - circle(at 50cm 50%)
PASS test unit (inline): cm - circle(at 50cm 50cm)
FAIL test unit (inline): cm - circle(at left 50cm) assert_equals: expected "circle(at 0% 50cm)" but got "circle(at left 50cm)"
FAIL test unit (inline): cm - circle(at 50cm top) assert_equals: expected "circle(at 50cm 0%)" but got "circle(at 50cm top)"
FAIL test unit (inline): cm - circle(at right 80cm) assert_equals: expected "circle(at 100% 80cm)" but got "circle(at right 80cm)"
FAIL test unit (inline): cm - circle(at 70cm bottom) assert_equals: expected "circle(at 70cm 100%)" but got "circle(at 70cm bottom)"
FAIL test unit (inline): cm - circle(at center 60cm) assert_equals: expected "circle(at 50% 60cm)" but got "circle(at center 60cm)"
FAIL test unit (inline): cm - circle(at 60cm center) assert_equals: expected "circle(at 60cm 50%)" but got "circle(at 60cm center)"
FAIL test unit (inline): cm - circle(at left 50% top 50cm) assert_equals: expected "circle(at 50% 50cm)" but got "circle(at left 50% top 50cm)"
FAIL test unit (inline): cm - circle(at left 50% bottom 70cm) assert_equals: expected "circle(at 50% calc(100% - 2645.67px))" but got "circle(at left 50% bottom 70cm)"
FAIL test unit (inline): cm - circle(at left 50cm top 50%) assert_equals: expected "circle(at 50cm 50%)" but got "circle(at left 50cm top 50%)"
FAIL test unit (inline): cm - circle(at left 50cm top 50cm) assert_equals: expected "circle(at 50cm 50cm)" but got "circle(at left 50cm top 50cm)"
FAIL test unit (inline): cm - circle(at left 50cm bottom 70%) assert_equals: expected "circle(at 50cm calc(30%))" but got "circle(at left 50cm bottom 70%)"
FAIL test unit (inline): cm - circle(at top 50% left 50cm) assert_equals: expected "circle(at 50cm 50%)" but got "circle(at left 50cm top 50%)"
FAIL test unit (inline): cm - circle(at top 50% right 80cm) assert_equals: expected "circle(at calc(100% - 3023.62px) 50%)" but got "circle(at right 80cm top 50%)"
FAIL test unit (inline): cm - circle(at top 50cm left 50%) assert_equals: expected "circle(at 50% 50cm)" but got "circle(at left 50% top 50cm)"
FAIL test unit (inline): cm - circle(at top 50cm left 50cm) assert_equals: expected "circle(at 50cm 50cm)" but got "circle(at left 50cm top 50cm)"
FAIL test unit (inline): cm - circle(at top 50cm right 80%) assert_equals: expected "circle(at calc(20%) 50cm)" but got "circle(at right 80% top 50cm)"
FAIL test unit (inline): cm - circle(at bottom 70% left 50cm) assert_equals: expected "circle(at 50cm calc(30%))" but got "circle(at left 50cm bottom 70%)"
FAIL test unit (inline): cm - circle(at bottom 70% right 80cm) assert_equals: expected "circle(at calc(100% - 3023.62px) calc(30%))" but got "circle(at right 80cm bottom 70%)"
FAIL test unit (inline): cm - circle(at bottom 70cm left 50%) assert_equals: expected "circle(at 50% calc(100% - 2645.67px))" but got "circle(at left 50% bottom 70cm)"
FAIL test unit (inline): cm - circle(at bottom 70cm right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 2645.67px))" but got "circle(at right 50% bottom 70cm)"
FAIL test unit (inline): cm - circle(at bottom 70cm right 80cm) assert_equals: expected "circle(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "circle(at right 80cm bottom 70cm)"
FAIL test unit (inline): cm - circle(at right 80% top 50cm) assert_equals: expected "circle(at calc(20%) 50cm)" but got "circle(at right 80% top 50cm)"
FAIL test unit (inline): cm - circle(at right 80% bottom 70cm) assert_equals: expected "circle(at calc(20%) calc(100% - 2645.67px))" but got "circle(at right 80% bottom 70cm)"
FAIL test unit (inline): cm - circle(at right 80cm top 50%) assert_equals: expected "circle(at calc(100% - 3023.62px) 50%)" but got "circle(at right 80cm top 50%)"
FAIL test unit (inline): cm - circle(at right 80cm bottom 70%) assert_equals: expected "circle(at calc(100% - 3023.62px) calc(30%))" but got "circle(at right 80cm bottom 70%)"
FAIL test unit (inline): cm - circle(at right 80cm bottom 70cm) assert_equals: expected "circle(at calc(100% - 3023.62px) calc(100% - 2645.67px))" but got "circle(at right 80cm bottom 70cm)"
FAIL test unit (inline): mm - circle(at 50mm) assert_equals: expected "circle(at 50mm 50%)" but got "circle(at 50mm center)"
PASS test unit (inline): mm - circle(at 50% 50mm)
PASS test unit (inline): mm - circle(at 50mm 50%)
PASS test unit (inline): mm - circle(at 50mm 50mm)
FAIL test unit (inline): mm - circle(at left 50mm) assert_equals: expected "circle(at 0% 50mm)" but got "circle(at left 50mm)"
FAIL test unit (inline): mm - circle(at 50mm top) assert_equals: expected "circle(at 50mm 0%)" but got "circle(at 50mm top)"
FAIL test unit (inline): mm - circle(at right 80mm) assert_equals: expected "circle(at 100% 80mm)" but got "circle(at right 80mm)"
FAIL test unit (inline): mm - circle(at 70mm bottom) assert_equals: expected "circle(at 70mm 100%)" but got "circle(at 70mm bottom)"
FAIL test unit (inline): mm - circle(at center 60mm) assert_equals: expected "circle(at 50% 60mm)" but got "circle(at center 60mm)"
FAIL test unit (inline): mm - circle(at 60mm center) assert_equals: expected "circle(at 60mm 50%)" but got "circle(at 60mm center)"
FAIL test unit (inline): mm - circle(at left 50% top 50mm) assert_equals: expected "circle(at 50% 50mm)" but got "circle(at left 50% top 50mm)"
FAIL test unit (inline): mm - circle(at left 50% bottom 70mm) assert_equals: expected "circle(at 50% calc(100% - 264.567px))" but got "circle(at left 50% bottom 70mm)"
FAIL test unit (inline): mm - circle(at left 50mm top 50%) assert_equals: expected "circle(at 50mm 50%)" but got "circle(at left 50mm top 50%)"
FAIL test unit (inline): mm - circle(at left 50mm top 50mm) assert_equals: expected "circle(at 50mm 50mm)" but got "circle(at left 50mm top 50mm)"
FAIL test unit (inline): mm - circle(at left 50mm bottom 70%) assert_equals: expected "circle(at 50mm calc(30%))" but got "circle(at left 50mm bottom 70%)"
FAIL test unit (inline): mm - circle(at top 50% left 50mm) assert_equals: expected "circle(at 50mm 50%)" but got "circle(at left 50mm top 50%)"
FAIL test unit (inline): mm - circle(at top 50% right 80mm) assert_equals: expected "circle(at calc(100% - 302.362px) 50%)" but got "circle(at right 80mm top 50%)"
FAIL test unit (inline): mm - circle(at top 50mm left 50%) assert_equals: expected "circle(at 50% 50mm)" but got "circle(at left 50% top 50mm)"
FAIL test unit (inline): mm - circle(at top 50mm left 50mm) assert_equals: expected "circle(at 50mm 50mm)" but got "circle(at left 50mm top 50mm)"
FAIL test unit (inline): mm - circle(at top 50mm right 80%) assert_equals: expected "circle(at calc(20%) 50mm)" but got "circle(at right 80% top 50mm)"
FAIL test unit (inline): mm - circle(at bottom 70% left 50mm) assert_equals: expected "circle(at 50mm calc(30%))" but got "circle(at left 50mm bottom 70%)"
FAIL test unit (inline): mm - circle(at bottom 70% right 80mm) assert_equals: expected "circle(at calc(100% - 302.362px) calc(30%))" but got "circle(at right 80mm bottom 70%)"
FAIL test unit (inline): mm - circle(at bottom 70mm left 50%) assert_equals: expected "circle(at 50% calc(100% - 264.567px))" but got "circle(at left 50% bottom 70mm)"
FAIL test unit (inline): mm - circle(at bottom 70mm right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 264.567px))" but got "circle(at right 50% bottom 70mm)"
FAIL test unit (inline): mm - circle(at bottom 70mm right 80mm) assert_equals: expected "circle(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "circle(at right 80mm bottom 70mm)"
FAIL test unit (inline): mm - circle(at right 80% top 50mm) assert_equals: expected "circle(at calc(20%) 50mm)" but got "circle(at right 80% top 50mm)"
FAIL test unit (inline): mm - circle(at right 80% bottom 70mm) assert_equals: expected "circle(at calc(20%) calc(100% - 264.567px))" but got "circle(at right 80% bottom 70mm)"
FAIL test unit (inline): mm - circle(at right 80mm top 50%) assert_equals: expected "circle(at calc(100% - 302.362px) 50%)" but got "circle(at right 80mm top 50%)"
FAIL test unit (inline): mm - circle(at right 80mm bottom 70%) assert_equals: expected "circle(at calc(100% - 302.362px) calc(30%))" but got "circle(at right 80mm bottom 70%)"
FAIL test unit (inline): mm - circle(at right 80mm bottom 70mm) assert_equals: expected "circle(at calc(100% - 302.362px) calc(100% - 264.567px))" but got "circle(at right 80mm bottom 70mm)"
FAIL test unit (inline): in - circle(at 50in) assert_equals: expected "circle(at 50in 50%)" but got "circle(at 50in center)"
PASS test unit (inline): in - circle(at 50% 50in)
PASS test unit (inline): in - circle(at 50in 50%)
PASS test unit (inline): in - circle(at 50in 50in)
FAIL test unit (inline): in - circle(at left 50in) assert_equals: expected "circle(at 0% 50in)" but got "circle(at left 50in)"
FAIL test unit (inline): in - circle(at 50in top) assert_equals: expected "circle(at 50in 0%)" but got "circle(at 50in top)"
FAIL test unit (inline): in - circle(at right 80in) assert_equals: expected "circle(at 100% 80in)" but got "circle(at right 80in)"
FAIL test unit (inline): in - circle(at 70in bottom) assert_equals: expected "circle(at 70in 100%)" but got "circle(at 70in bottom)"
FAIL test unit (inline): in - circle(at center 60in) assert_equals: expected "circle(at 50% 60in)" but got "circle(at center 60in)"
FAIL test unit (inline): in - circle(at 60in center) assert_equals: expected "circle(at 60in 50%)" but got "circle(at 60in center)"
FAIL test unit (inline): in - circle(at left 50% top 50in) assert_equals: expected "circle(at 50% 50in)" but got "circle(at left 50% top 50in)"
FAIL test unit (inline): in - circle(at left 50% bottom 70in) assert_equals: expected "circle(at 50% calc(100% - 6720px))" but got "circle(at left 50% bottom 70in)"
FAIL test unit (inline): in - circle(at left 50in top 50%) assert_equals: expected "circle(at 50in 50%)" but got "circle(at left 50in top 50%)"
FAIL test unit (inline): in - circle(at left 50in top 50in) assert_equals: expected "circle(at 50in 50in)" but got "circle(at left 50in top 50in)"
FAIL test unit (inline): in - circle(at left 50in bottom 70%) assert_equals: expected "circle(at 50in calc(30%))" but got "circle(at left 50in bottom 70%)"
FAIL test unit (inline): in - circle(at top 50% left 50in) assert_equals: expected "circle(at 50in 50%)" but got "circle(at left 50in top 50%)"
FAIL test unit (inline): in - circle(at top 50% right 80in) assert_equals: expected "circle(at calc(100% - 7680px) 50%)" but got "circle(at right 80in top 50%)"
FAIL test unit (inline): in - circle(at top 50in left 50%) assert_equals: expected "circle(at 50% 50in)" but got "circle(at left 50% top 50in)"
FAIL test unit (inline): in - circle(at top 50in left 50in) assert_equals: expected "circle(at 50in 50in)" but got "circle(at left 50in top 50in)"
FAIL test unit (inline): in - circle(at top 50in right 80%) assert_equals: expected "circle(at calc(20%) 50in)" but got "circle(at right 80% top 50in)"
FAIL test unit (inline): in - circle(at bottom 70% left 50in) assert_equals: expected "circle(at 50in calc(30%))" but got "circle(at left 50in bottom 70%)"
FAIL test unit (inline): in - circle(at bottom 70% right 80in) assert_equals: expected "circle(at calc(100% - 7680px) calc(30%))" but got "circle(at right 80in bottom 70%)"
FAIL test unit (inline): in - circle(at bottom 70in left 50%) assert_equals: expected "circle(at 50% calc(100% - 6720px))" but got "circle(at left 50% bottom 70in)"
FAIL test unit (inline): in - circle(at bottom 70in right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 6720px))" but got "circle(at right 50% bottom 70in)"
FAIL test unit (inline): in - circle(at bottom 70in right 80in) assert_equals: expected "circle(at calc(100% - 7680px) calc(100% - 6720px))" but got "circle(at right 80in bottom 70in)"
FAIL test unit (inline): in - circle(at right 80% top 50in) assert_equals: expected "circle(at calc(20%) 50in)" but got "circle(at right 80% top 50in)"
FAIL test unit (inline): in - circle(at right 80% bottom 70in) assert_equals: expected "circle(at calc(20%) calc(100% - 6720px))" but got "circle(at right 80% bottom 70in)"
FAIL test unit (inline): in - circle(at right 80in top 50%) assert_equals: expected "circle(at calc(100% - 7680px) 50%)" but got "circle(at right 80in top 50%)"
FAIL test unit (inline): in - circle(at right 80in bottom 70%) assert_equals: expected "circle(at calc(100% - 7680px) calc(30%))" but got "circle(at right 80in bottom 70%)"
FAIL test unit (inline): in - circle(at right 80in bottom 70in) assert_equals: expected "circle(at calc(100% - 7680px) calc(100% - 6720px))" but got "circle(at right 80in bottom 70in)"
FAIL test unit (inline): pt - circle(at 50pt) assert_equals: expected "circle(at 50pt 50%)" but got "circle(at 50pt center)"
PASS test unit (inline): pt - circle(at 50% 50pt)
PASS test unit (inline): pt - circle(at 50pt 50%)
PASS test unit (inline): pt - circle(at 50pt 50pt)
FAIL test unit (inline): pt - circle(at left 50pt) assert_equals: expected "circle(at 0% 50pt)" but got "circle(at left 50pt)"
FAIL test unit (inline): pt - circle(at 50pt top) assert_equals: expected "circle(at 50pt 0%)" but got "circle(at 50pt top)"
FAIL test unit (inline): pt - circle(at right 80pt) assert_equals: expected "circle(at 100% 80pt)" but got "circle(at right 80pt)"
FAIL test unit (inline): pt - circle(at 70pt bottom) assert_equals: expected "circle(at 70pt 100%)" but got "circle(at 70pt bottom)"
FAIL test unit (inline): pt - circle(at center 60pt) assert_equals: expected "circle(at 50% 60pt)" but got "circle(at center 60pt)"
FAIL test unit (inline): pt - circle(at 60pt center) assert_equals: expected "circle(at 60pt 50%)" but got "circle(at 60pt center)"
FAIL test unit (inline): pt - circle(at left 50% top 50pt) assert_equals: expected "circle(at 50% 50pt)" but got "circle(at left 50% top 50pt)"
FAIL test unit (inline): pt - circle(at left 50% bottom 70pt) assert_equals: expected "circle(at 50% calc(100% - 93.3333px))" but got "circle(at left 50% bottom 70pt)"
FAIL test unit (inline): pt - circle(at left 50pt top 50%) assert_equals: expected "circle(at 50pt 50%)" but got "circle(at left 50pt top 50%)"
FAIL test unit (inline): pt - circle(at left 50pt top 50pt) assert_equals: expected "circle(at 50pt 50pt)" but got "circle(at left 50pt top 50pt)"
FAIL test unit (inline): pt - circle(at left 50pt bottom 70%) assert_equals: expected "circle(at 50pt calc(30%))" but got "circle(at left 50pt bottom 70%)"
FAIL test unit (inline): pt - circle(at top 50% left 50pt) assert_equals: expected "circle(at 50pt 50%)" but got "circle(at left 50pt top 50%)"
FAIL test unit (inline): pt - circle(at top 50% right 80pt) assert_equals: expected "circle(at calc(100% - 106.667px) 50%)" but got "circle(at right 80pt top 50%)"
FAIL test unit (inline): pt - circle(at top 50pt left 50%) assert_equals: expected "circle(at 50% 50pt)" but got "circle(at left 50% top 50pt)"
FAIL test unit (inline): pt - circle(at top 50pt left 50pt) assert_equals: expected "circle(at 50pt 50pt)" but got "circle(at left 50pt top 50pt)"
FAIL test unit (inline): pt - circle(at top 50pt right 80%) assert_equals: expected "circle(at calc(20%) 50pt)" but got "circle(at right 80% top 50pt)"
FAIL test unit (inline): pt - circle(at bottom 70% left 50pt) assert_equals: expected "circle(at 50pt calc(30%))" but got "circle(at left 50pt bottom 70%)"
FAIL test unit (inline): pt - circle(at bottom 70% right 80pt) assert_equals: expected "circle(at calc(100% - 106.667px) calc(30%))" but got "circle(at right 80pt bottom 70%)"
FAIL test unit (inline): pt - circle(at bottom 70pt left 50%) assert_equals: expected "circle(at 50% calc(100% - 93.3333px))" but got "circle(at left 50% bottom 70pt)"
FAIL test unit (inline): pt - circle(at bottom 70pt right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 93.3333px))" but got "circle(at right 50% bottom 70pt)"
FAIL test unit (inline): pt - circle(at bottom 70pt right 80pt) assert_equals: expected "circle(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "circle(at right 80pt bottom 70pt)"
FAIL test unit (inline): pt - circle(at right 80% top 50pt) assert_equals: expected "circle(at calc(20%) 50pt)" but got "circle(at right 80% top 50pt)"
FAIL test unit (inline): pt - circle(at right 80% bottom 70pt) assert_equals: expected "circle(at calc(20%) calc(100% - 93.3333px))" but got "circle(at right 80% bottom 70pt)"
FAIL test unit (inline): pt - circle(at right 80pt top 50%) assert_equals: expected "circle(at calc(100% - 106.667px) 50%)" but got "circle(at right 80pt top 50%)"
FAIL test unit (inline): pt - circle(at right 80pt bottom 70%) assert_equals: expected "circle(at calc(100% - 106.667px) calc(30%))" but got "circle(at right 80pt bottom 70%)"
FAIL test unit (inline): pt - circle(at right 80pt bottom 70pt) assert_equals: expected "circle(at calc(100% - 106.667px) calc(100% - 93.3333px))" but got "circle(at right 80pt bottom 70pt)"
FAIL test unit (inline): pc - circle(at 50pc) assert_equals: expected "circle(at 50pc 50%)" but got "circle(at 50pc center)"
PASS test unit (inline): pc - circle(at 50% 50pc)
PASS test unit (inline): pc - circle(at 50pc 50%)
PASS test unit (inline): pc - circle(at 50pc 50pc)
FAIL test unit (inline): pc - circle(at left 50pc) assert_equals: expected "circle(at 0% 50pc)" but got "circle(at left 50pc)"
FAIL test unit (inline): pc - circle(at 50pc top) assert_equals: expected "circle(at 50pc 0%)" but got "circle(at 50pc top)"
FAIL test unit (inline): pc - circle(at right 80pc) assert_equals: expected "circle(at 100% 80pc)" but got "circle(at right 80pc)"
FAIL test unit (inline): pc - circle(at 70pc bottom) assert_equals: expected "circle(at 70pc 100%)" but got "circle(at 70pc bottom)"
FAIL test unit (inline): pc - circle(at center 60pc) assert_equals: expected "circle(at 50% 60pc)" but got "circle(at center 60pc)"
FAIL test unit (inline): pc - circle(at 60pc center) assert_equals: expected "circle(at 60pc 50%)" but got "circle(at 60pc center)"
FAIL test unit (inline): pc - circle(at left 50% top 50pc) assert_equals: expected "circle(at 50% 50pc)" but got "circle(at left 50% top 50pc)"
FAIL test unit (inline): pc - circle(at left 50% bottom 70pc) assert_equals: expected "circle(at 50% calc(100% - 1120px))" but got "circle(at left 50% bottom 70pc)"
FAIL test unit (inline): pc - circle(at left 50pc top 50%) assert_equals: expected "circle(at 50pc 50%)" but got "circle(at left 50pc top 50%)"
FAIL test unit (inline): pc - circle(at left 50pc top 50pc) assert_equals: expected "circle(at 50pc 50pc)" but got "circle(at left 50pc top 50pc)"
FAIL test unit (inline): pc - circle(at left 50pc bottom 70%) assert_equals: expected "circle(at 50pc calc(30%))" but got "circle(at left 50pc bottom 70%)"
FAIL test unit (inline): pc - circle(at top 50% left 50pc) assert_equals: expected "circle(at 50pc 50%)" but got "circle(at left 50pc top 50%)"
FAIL test unit (inline): pc - circle(at top 50% right 80pc) assert_equals: expected "circle(at calc(100% - 1280px) 50%)" but got "circle(at right 80pc top 50%)"
FAIL test unit (inline): pc - circle(at top 50pc left 50%) assert_equals: expected "circle(at 50% 50pc)" but got "circle(at left 50% top 50pc)"
FAIL test unit (inline): pc - circle(at top 50pc left 50pc) assert_equals: expected "circle(at 50pc 50pc)" but got "circle(at left 50pc top 50pc)"
FAIL test unit (inline): pc - circle(at top 50pc right 80%) assert_equals: expected "circle(at calc(20%) 50pc)" but got "circle(at right 80% top 50pc)"
FAIL test unit (inline): pc - circle(at bottom 70% left 50pc) assert_equals: expected "circle(at 50pc calc(30%))" but got "circle(at left 50pc bottom 70%)"
FAIL test unit (inline): pc - circle(at bottom 70% right 80pc) assert_equals: expected "circle(at calc(100% - 1280px) calc(30%))" but got "circle(at right 80pc bottom 70%)"
FAIL test unit (inline): pc - circle(at bottom 70pc left 50%) assert_equals: expected "circle(at 50% calc(100% - 1120px))" but got "circle(at left 50% bottom 70pc)"
FAIL test unit (inline): pc - circle(at bottom 70pc right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 1120px))" but got "circle(at right 50% bottom 70pc)"
FAIL test unit (inline): pc - circle(at bottom 70pc right 80pc) assert_equals: expected "circle(at calc(100% - 1280px) calc(100% - 1120px))" but got "circle(at right 80pc bottom 70pc)"
FAIL test unit (inline): pc - circle(at right 80% top 50pc) assert_equals: expected "circle(at calc(20%) 50pc)" but got "circle(at right 80% top 50pc)"
FAIL test unit (inline): pc - circle(at right 80% bottom 70pc) assert_equals: expected "circle(at calc(20%) calc(100% - 1120px))" but got "circle(at right 80% bottom 70pc)"
FAIL test unit (inline): pc - circle(at right 80pc top 50%) assert_equals: expected "circle(at calc(100% - 1280px) 50%)" but got "circle(at right 80pc top 50%)"
FAIL test unit (inline): pc - circle(at right 80pc bottom 70%) assert_equals: expected "circle(at calc(100% - 1280px) calc(30%))" but got "circle(at right 80pc bottom 70%)"
FAIL test unit (inline): pc - circle(at right 80pc bottom 70pc) assert_equals: expected "circle(at calc(100% - 1280px) calc(100% - 1120px))" but got "circle(at right 80pc bottom 70pc)"
FAIL test unit (inline): em - circle(at 50em) assert_equals: expected "circle(at 50em 50%)" but got "circle(at 50em center)"
PASS test unit (inline): em - circle(at 50% 50em)
PASS test unit (inline): em - circle(at 50em 50%)
PASS test unit (inline): em - circle(at 50em 50em)
FAIL test unit (inline): em - circle(at left 50em) assert_equals: expected "circle(at 0% 50em)" but got "circle(at left 50em)"
FAIL test unit (inline): em - circle(at 50em top) assert_equals: expected "circle(at 50em 0%)" but got "circle(at 50em top)"
FAIL test unit (inline): em - circle(at right 80em) assert_equals: expected "circle(at 100% 80em)" but got "circle(at right 80em)"
FAIL test unit (inline): em - circle(at 70em bottom) assert_equals: expected "circle(at 70em 100%)" but got "circle(at 70em bottom)"
FAIL test unit (inline): em - circle(at center 60em) assert_equals: expected "circle(at 50% 60em)" but got "circle(at center 60em)"
FAIL test unit (inline): em - circle(at 60em center) assert_equals: expected "circle(at 60em 50%)" but got "circle(at 60em center)"
FAIL test unit (inline): em - circle(at left 50% top 50em) assert_equals: expected "circle(at 50% 50em)" but got "circle(at left 50% top 50em)"
FAIL test unit (inline): em - circle(at left 50% bottom 70em) assert_equals: expected "circle(at 50% calc(100% - 70em))" but got "circle(at left 50% bottom 70em)"
FAIL test unit (inline): em - circle(at left 50em top 50%) assert_equals: expected "circle(at 50em 50%)" but got "circle(at left 50em top 50%)"
FAIL test unit (inline): em - circle(at left 50em top 50em) assert_equals: expected "circle(at 50em 50em)" but got "circle(at left 50em top 50em)"
FAIL test unit (inline): em - circle(at left 50em bottom 70%) assert_equals: expected "circle(at 50em calc(30%))" but got "circle(at left 50em bottom 70%)"
FAIL test unit (inline): em - circle(at top 50% left 50em) assert_equals: expected "circle(at 50em 50%)" but got "circle(at left 50em top 50%)"
FAIL test unit (inline): em - circle(at top 50% right 80em) assert_equals: expected "circle(at calc(100% - 80em) 50%)" but got "circle(at right 80em top 50%)"
FAIL test unit (inline): em - circle(at top 50em left 50%) assert_equals: expected "circle(at 50% 50em)" but got "circle(at left 50% top 50em)"
FAIL test unit (inline): em - circle(at top 50em left 50em) assert_equals: expected "circle(at 50em 50em)" but got "circle(at left 50em top 50em)"
FAIL test unit (inline): em - circle(at top 50em right 80%) assert_equals: expected "circle(at calc(20%) 50em)" but got "circle(at right 80% top 50em)"
FAIL test unit (inline): em - circle(at bottom 70% left 50em) assert_equals: expected "circle(at 50em calc(30%))" but got "circle(at left 50em bottom 70%)"
FAIL test unit (inline): em - circle(at bottom 70% right 80em) assert_equals: expected "circle(at calc(100% - 80em) calc(30%))" but got "circle(at right 80em bottom 70%)"
FAIL test unit (inline): em - circle(at bottom 70em left 50%) assert_equals: expected "circle(at 50% calc(100% - 70em))" but got "circle(at left 50% bottom 70em)"
FAIL test unit (inline): em - circle(at bottom 70em right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70em))" but got "circle(at right 50% bottom 70em)"
FAIL test unit (inline): em - circle(at bottom 70em right 80em) assert_equals: expected "circle(at calc(100% - 80em) calc(100% - 70em))" but got "circle(at right 80em bottom 70em)"
FAIL test unit (inline): em - circle(at right 80% top 50em) assert_equals: expected "circle(at calc(20%) 50em)" but got "circle(at right 80% top 50em)"
FAIL test unit (inline): em - circle(at right 80% bottom 70em) assert_equals: expected "circle(at calc(20%) calc(100% - 70em))" but got "circle(at right 80% bottom 70em)"
FAIL test unit (inline): em - circle(at right 80em top 50%) assert_equals: expected "circle(at calc(100% - 80em) 50%)" but got "circle(at right 80em top 50%)"
FAIL test unit (inline): em - circle(at right 80em bottom 70%) assert_equals: expected "circle(at calc(100% - 80em) calc(30%))" but got "circle(at right 80em bottom 70%)"
FAIL test unit (inline): em - circle(at right 80em bottom 70em) assert_equals: expected "circle(at calc(100% - 80em) calc(100% - 70em))" but got "circle(at right 80em bottom 70em)"
FAIL test unit (inline): ex - circle(at 50ex) assert_equals: expected "circle(at 50ex 50%)" but got "circle(at 50ex center)"
PASS test unit (inline): ex - circle(at 50% 50ex)
PASS test unit (inline): ex - circle(at 50ex 50%)
PASS test unit (inline): ex - circle(at 50ex 50ex)
FAIL test unit (inline): ex - circle(at left 50ex) assert_equals: expected "circle(at 0% 50ex)" but got "circle(at left 50ex)"
FAIL test unit (inline): ex - circle(at 50ex top) assert_equals: expected "circle(at 50ex 0%)" but got "circle(at 50ex top)"
FAIL test unit (inline): ex - circle(at right 80ex) assert_equals: expected "circle(at 100% 80ex)" but got "circle(at right 80ex)"
FAIL test unit (inline): ex - circle(at 70ex bottom) assert_equals: expected "circle(at 70ex 100%)" but got "circle(at 70ex bottom)"
FAIL test unit (inline): ex - circle(at center 60ex) assert_equals: expected "circle(at 50% 60ex)" but got "circle(at center 60ex)"
FAIL test unit (inline): ex - circle(at 60ex center) assert_equals: expected "circle(at 60ex 50%)" but got "circle(at 60ex center)"
FAIL test unit (inline): ex - circle(at left 50% top 50ex) assert_equals: expected "circle(at 50% 50ex)" but got "circle(at left 50% top 50ex)"
FAIL test unit (inline): ex - circle(at left 50% bottom 70ex) assert_equals: expected "circle(at 50% calc(100% - 70ex))" but got "circle(at left 50% bottom 70ex)"
FAIL test unit (inline): ex - circle(at left 50ex top 50%) assert_equals: expected "circle(at 50ex 50%)" but got "circle(at left 50ex top 50%)"
FAIL test unit (inline): ex - circle(at left 50ex top 50ex) assert_equals: expected "circle(at 50ex 50ex)" but got "circle(at left 50ex top 50ex)"
FAIL test unit (inline): ex - circle(at left 50ex bottom 70%) assert_equals: expected "circle(at 50ex calc(30%))" but got "circle(at left 50ex bottom 70%)"
FAIL test unit (inline): ex - circle(at top 50% left 50ex) assert_equals: expected "circle(at 50ex 50%)" but got "circle(at left 50ex top 50%)"
FAIL test unit (inline): ex - circle(at top 50% right 80ex) assert_equals: expected "circle(at calc(100% - 80ex) 50%)" but got "circle(at right 80ex top 50%)"
FAIL test unit (inline): ex - circle(at top 50ex left 50%) assert_equals: expected "circle(at 50% 50ex)" but got "circle(at left 50% top 50ex)"
FAIL test unit (inline): ex - circle(at top 50ex left 50ex) assert_equals: expected "circle(at 50ex 50ex)" but got "circle(at left 50ex top 50ex)"
FAIL test unit (inline): ex - circle(at top 50ex right 80%) assert_equals: expected "circle(at calc(20%) 50ex)" but got "circle(at right 80% top 50ex)"
FAIL test unit (inline): ex - circle(at bottom 70% left 50ex) assert_equals: expected "circle(at 50ex calc(30%))" but got "circle(at left 50ex bottom 70%)"
FAIL test unit (inline): ex - circle(at bottom 70% right 80ex) assert_equals: expected "circle(at calc(100% - 80ex) calc(30%))" but got "circle(at right 80ex bottom 70%)"
FAIL test unit (inline): ex - circle(at bottom 70ex left 50%) assert_equals: expected "circle(at 50% calc(100% - 70ex))" but got "circle(at left 50% bottom 70ex)"
FAIL test unit (inline): ex - circle(at bottom 70ex right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70ex))" but got "circle(at right 50% bottom 70ex)"
FAIL test unit (inline): ex - circle(at bottom 70ex right 80ex) assert_equals: expected "circle(at calc(100% - 80ex) calc(100% - 70ex))" but got "circle(at right 80ex bottom 70ex)"
FAIL test unit (inline): ex - circle(at right 80% top 50ex) assert_equals: expected "circle(at calc(20%) 50ex)" but got "circle(at right 80% top 50ex)"
FAIL test unit (inline): ex - circle(at right 80% bottom 70ex) assert_equals: expected "circle(at calc(20%) calc(100% - 70ex))" but got "circle(at right 80% bottom 70ex)"
FAIL test unit (inline): ex - circle(at right 80ex top 50%) assert_equals: expected "circle(at calc(100% - 80ex) 50%)" but got "circle(at right 80ex top 50%)"
FAIL test unit (inline): ex - circle(at right 80ex bottom 70%) assert_equals: expected "circle(at calc(100% - 80ex) calc(30%))" but got "circle(at right 80ex bottom 70%)"
FAIL test unit (inline): ex - circle(at right 80ex bottom 70ex) assert_equals: expected "circle(at calc(100% - 80ex) calc(100% - 70ex))" but got "circle(at right 80ex bottom 70ex)"
FAIL test unit (inline): ch - circle(at 50ch) assert_equals: expected "circle(at 50ch 50%)" but got "circle(at 50ch center)"
PASS test unit (inline): ch - circle(at 50% 50ch)
PASS test unit (inline): ch - circle(at 50ch 50%)
PASS test unit (inline): ch - circle(at 50ch 50ch)
FAIL test unit (inline): ch - circle(at left 50ch) assert_equals: expected "circle(at 0% 50ch)" but got "circle(at left 50ch)"
FAIL test unit (inline): ch - circle(at 50ch top) assert_equals: expected "circle(at 50ch 0%)" but got "circle(at 50ch top)"
FAIL test unit (inline): ch - circle(at right 80ch) assert_equals: expected "circle(at 100% 80ch)" but got "circle(at right 80ch)"
FAIL test unit (inline): ch - circle(at 70ch bottom) assert_equals: expected "circle(at 70ch 100%)" but got "circle(at 70ch bottom)"
FAIL test unit (inline): ch - circle(at center 60ch) assert_equals: expected "circle(at 50% 60ch)" but got "circle(at center 60ch)"
FAIL test unit (inline): ch - circle(at 60ch center) assert_equals: expected "circle(at 60ch 50%)" but got "circle(at 60ch center)"
FAIL test unit (inline): ch - circle(at left 50% top 50ch) assert_equals: expected "circle(at 50% 50ch)" but got "circle(at left 50% top 50ch)"
FAIL test unit (inline): ch - circle(at left 50% bottom 70ch) assert_equals: expected "circle(at 50% calc(100% - 70ch))" but got "circle(at left 50% bottom 70ch)"
FAIL test unit (inline): ch - circle(at left 50ch top 50%) assert_equals: expected "circle(at 50ch 50%)" but got "circle(at left 50ch top 50%)"
FAIL test unit (inline): ch - circle(at left 50ch top 50ch) assert_equals: expected "circle(at 50ch 50ch)" but got "circle(at left 50ch top 50ch)"
FAIL test unit (inline): ch - circle(at left 50ch bottom 70%) assert_equals: expected "circle(at 50ch calc(30%))" but got "circle(at left 50ch bottom 70%)"
FAIL test unit (inline): ch - circle(at top 50% left 50ch) assert_equals: expected "circle(at 50ch 50%)" but got "circle(at left 50ch top 50%)"
FAIL test unit (inline): ch - circle(at top 50% right 80ch) assert_equals: expected "circle(at calc(100% - 80ch) 50%)" but got "circle(at right 80ch top 50%)"
FAIL test unit (inline): ch - circle(at top 50ch left 50%) assert_equals: expected "circle(at 50% 50ch)" but got "circle(at left 50% top 50ch)"
FAIL test unit (inline): ch - circle(at top 50ch left 50ch) assert_equals: expected "circle(at 50ch 50ch)" but got "circle(at left 50ch top 50ch)"
FAIL test unit (inline): ch - circle(at top 50ch right 80%) assert_equals: expected "circle(at calc(20%) 50ch)" but got "circle(at right 80% top 50ch)"
FAIL test unit (inline): ch - circle(at bottom 70% left 50ch) assert_equals: expected "circle(at 50ch calc(30%))" but got "circle(at left 50ch bottom 70%)"
FAIL test unit (inline): ch - circle(at bottom 70% right 80ch) assert_equals: expected "circle(at calc(100% - 80ch) calc(30%))" but got "circle(at right 80ch bottom 70%)"
FAIL test unit (inline): ch - circle(at bottom 70ch left 50%) assert_equals: expected "circle(at 50% calc(100% - 70ch))" but got "circle(at left 50% bottom 70ch)"
FAIL test unit (inline): ch - circle(at bottom 70ch right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70ch))" but got "circle(at right 50% bottom 70ch)"
FAIL test unit (inline): ch - circle(at bottom 70ch right 80ch) assert_equals: expected "circle(at calc(100% - 80ch) calc(100% - 70ch))" but got "circle(at right 80ch bottom 70ch)"
FAIL test unit (inline): ch - circle(at right 80% top 50ch) assert_equals: expected "circle(at calc(20%) 50ch)" but got "circle(at right 80% top 50ch)"
FAIL test unit (inline): ch - circle(at right 80% bottom 70ch) assert_equals: expected "circle(at calc(20%) calc(100% - 70ch))" but got "circle(at right 80% bottom 70ch)"
FAIL test unit (inline): ch - circle(at right 80ch top 50%) assert_equals: expected "circle(at calc(100% - 80ch) 50%)" but got "circle(at right 80ch top 50%)"
FAIL test unit (inline): ch - circle(at right 80ch bottom 70%) assert_equals: expected "circle(at calc(100% - 80ch) calc(30%))" but got "circle(at right 80ch bottom 70%)"
FAIL test unit (inline): ch - circle(at right 80ch bottom 70ch) assert_equals: expected "circle(at calc(100% - 80ch) calc(100% - 70ch))" but got "circle(at right 80ch bottom 70ch)"
FAIL test unit (inline): rem - circle(at 50rem) assert_equals: expected "circle(at 50rem 50%)" but got "circle(at 50rem center)"
PASS test unit (inline): rem - circle(at 50% 50rem)
PASS test unit (inline): rem - circle(at 50rem 50%)
PASS test unit (inline): rem - circle(at 50rem 50rem)
FAIL test unit (inline): rem - circle(at left 50rem) assert_equals: expected "circle(at 0% 50rem)" but got "circle(at left 50rem)"
FAIL test unit (inline): rem - circle(at 50rem top) assert_equals: expected "circle(at 50rem 0%)" but got "circle(at 50rem top)"
FAIL test unit (inline): rem - circle(at right 80rem) assert_equals: expected "circle(at 100% 80rem)" but got "circle(at right 80rem)"
FAIL test unit (inline): rem - circle(at 70rem bottom) assert_equals: expected "circle(at 70rem 100%)" but got "circle(at 70rem bottom)"
FAIL test unit (inline): rem - circle(at center 60rem) assert_equals: expected "circle(at 50% 60rem)" but got "circle(at center 60rem)"
FAIL test unit (inline): rem - circle(at 60rem center) assert_equals: expected "circle(at 60rem 50%)" but got "circle(at 60rem center)"
FAIL test unit (inline): rem - circle(at left 50% top 50rem) assert_equals: expected "circle(at 50% 50rem)" but got "circle(at left 50% top 50rem)"
FAIL test unit (inline): rem - circle(at left 50% bottom 70rem) assert_equals: expected "circle(at 50% calc(100% - 70rem))" but got "circle(at left 50% bottom 70rem)"
FAIL test unit (inline): rem - circle(at left 50rem top 50%) assert_equals: expected "circle(at 50rem 50%)" but got "circle(at left 50rem top 50%)"
FAIL test unit (inline): rem - circle(at left 50rem top 50rem) assert_equals: expected "circle(at 50rem 50rem)" but got "circle(at left 50rem top 50rem)"
FAIL test unit (inline): rem - circle(at left 50rem bottom 70%) assert_equals: expected "circle(at 50rem calc(30%))" but got "circle(at left 50rem bottom 70%)"
FAIL test unit (inline): rem - circle(at top 50% left 50rem) assert_equals: expected "circle(at 50rem 50%)" but got "circle(at left 50rem top 50%)"
FAIL test unit (inline): rem - circle(at top 50% right 80rem) assert_equals: expected "circle(at calc(100% - 80rem) 50%)" but got "circle(at right 80rem top 50%)"
FAIL test unit (inline): rem - circle(at top 50rem left 50%) assert_equals: expected "circle(at 50% 50rem)" but got "circle(at left 50% top 50rem)"
FAIL test unit (inline): rem - circle(at top 50rem left 50rem) assert_equals: expected "circle(at 50rem 50rem)" but got "circle(at left 50rem top 50rem)"
FAIL test unit (inline): rem - circle(at top 50rem right 80%) assert_equals: expected "circle(at calc(20%) 50rem)" but got "circle(at right 80% top 50rem)"
FAIL test unit (inline): rem - circle(at bottom 70% left 50rem) assert_equals: expected "circle(at 50rem calc(30%))" but got "circle(at left 50rem bottom 70%)"
FAIL test unit (inline): rem - circle(at bottom 70% right 80rem) assert_equals: expected "circle(at calc(100% - 80rem) calc(30%))" but got "circle(at right 80rem bottom 70%)"
FAIL test unit (inline): rem - circle(at bottom 70rem left 50%) assert_equals: expected "circle(at 50% calc(100% - 70rem))" but got "circle(at left 50% bottom 70rem)"
FAIL test unit (inline): rem - circle(at bottom 70rem right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70rem))" but got "circle(at right 50% bottom 70rem)"
FAIL test unit (inline): rem - circle(at bottom 70rem right 80rem) assert_equals: expected "circle(at calc(100% - 80rem) calc(100% - 70rem))" but got "circle(at right 80rem bottom 70rem)"
FAIL test unit (inline): rem - circle(at right 80% top 50rem) assert_equals: expected "circle(at calc(20%) 50rem)" but got "circle(at right 80% top 50rem)"
FAIL test unit (inline): rem - circle(at right 80% bottom 70rem) assert_equals: expected "circle(at calc(20%) calc(100% - 70rem))" but got "circle(at right 80% bottom 70rem)"
FAIL test unit (inline): rem - circle(at right 80rem top 50%) assert_equals: expected "circle(at calc(100% - 80rem) 50%)" but got "circle(at right 80rem top 50%)"
FAIL test unit (inline): rem - circle(at right 80rem bottom 70%) assert_equals: expected "circle(at calc(100% - 80rem) calc(30%))" but got "circle(at right 80rem bottom 70%)"
FAIL test unit (inline): rem - circle(at right 80rem bottom 70rem) assert_equals: expected "circle(at calc(100% - 80rem) calc(100% - 70rem))" but got "circle(at right 80rem bottom 70rem)"
FAIL test unit (inline): vw - circle(at 50vw) assert_equals: expected "circle(at 50vw 50%)" but got "circle(at 50vw center)"
PASS test unit (inline): vw - circle(at 50% 50vw)
PASS test unit (inline): vw - circle(at 50vw 50%)
PASS test unit (inline): vw - circle(at 50vw 50vw)
FAIL test unit (inline): vw - circle(at left 50vw) assert_equals: expected "circle(at 0% 50vw)" but got "circle(at left 50vw)"
FAIL test unit (inline): vw - circle(at 50vw top) assert_equals: expected "circle(at 50vw 0%)" but got "circle(at 50vw top)"
FAIL test unit (inline): vw - circle(at right 80vw) assert_equals: expected "circle(at 100% 80vw)" but got "circle(at right 80vw)"
FAIL test unit (inline): vw - circle(at 70vw bottom) assert_equals: expected "circle(at 70vw 100%)" but got "circle(at 70vw bottom)"
FAIL test unit (inline): vw - circle(at center 60vw) assert_equals: expected "circle(at 50% 60vw)" but got "circle(at center 60vw)"
FAIL test unit (inline): vw - circle(at 60vw center) assert_equals: expected "circle(at 60vw 50%)" but got "circle(at 60vw center)"
FAIL test unit (inline): vw - circle(at left 50% top 50vw) assert_equals: expected "circle(at 50% 50vw)" but got "circle(at left 50% top 50vw)"
FAIL test unit (inline): vw - circle(at left 50% bottom 70vw) assert_equals: expected "circle(at 50% calc(100% - 70vw))" but got "circle(at left 50% bottom 70vw)"
FAIL test unit (inline): vw - circle(at left 50vw top 50%) assert_equals: expected "circle(at 50vw 50%)" but got "circle(at left 50vw top 50%)"
FAIL test unit (inline): vw - circle(at left 50vw top 50vw) assert_equals: expected "circle(at 50vw 50vw)" but got "circle(at left 50vw top 50vw)"
FAIL test unit (inline): vw - circle(at left 50vw bottom 70%) assert_equals: expected "circle(at 50vw calc(30%))" but got "circle(at left 50vw bottom 70%)"
FAIL test unit (inline): vw - circle(at top 50% left 50vw) assert_equals: expected "circle(at 50vw 50%)" but got "circle(at left 50vw top 50%)"
FAIL test unit (inline): vw - circle(at top 50% right 80vw) assert_equals: expected "circle(at calc(100% - 80vw) 50%)" but got "circle(at right 80vw top 50%)"
FAIL test unit (inline): vw - circle(at top 50vw left 50%) assert_equals: expected "circle(at 50% 50vw)" but got "circle(at left 50% top 50vw)"
FAIL test unit (inline): vw - circle(at top 50vw left 50vw) assert_equals: expected "circle(at 50vw 50vw)" but got "circle(at left 50vw top 50vw)"
FAIL test unit (inline): vw - circle(at top 50vw right 80%) assert_equals: expected "circle(at calc(20%) 50vw)" but got "circle(at right 80% top 50vw)"
FAIL test unit (inline): vw - circle(at bottom 70% left 50vw) assert_equals: expected "circle(at 50vw calc(30%))" but got "circle(at left 50vw bottom 70%)"
FAIL test unit (inline): vw - circle(at bottom 70% right 80vw) assert_equals: expected "circle(at calc(100% - 80vw) calc(30%))" but got "circle(at right 80vw bottom 70%)"
FAIL test unit (inline): vw - circle(at bottom 70vw left 50%) assert_equals: expected "circle(at 50% calc(100% - 70vw))" but got "circle(at left 50% bottom 70vw)"
FAIL test unit (inline): vw - circle(at bottom 70vw right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70vw))" but got "circle(at right 50% bottom 70vw)"
FAIL test unit (inline): vw - circle(at bottom 70vw right 80vw) assert_equals: expected "circle(at calc(100% - 80vw) calc(100% - 70vw))" but got "circle(at right 80vw bottom 70vw)"
FAIL test unit (inline): vw - circle(at right 80% top 50vw) assert_equals: expected "circle(at calc(20%) 50vw)" but got "circle(at right 80% top 50vw)"
FAIL test unit (inline): vw - circle(at right 80% bottom 70vw) assert_equals: expected "circle(at calc(20%) calc(100% - 70vw))" but got "circle(at right 80% bottom 70vw)"
FAIL test unit (inline): vw - circle(at right 80vw top 50%) assert_equals: expected "circle(at calc(100% - 80vw) 50%)" but got "circle(at right 80vw top 50%)"
FAIL test unit (inline): vw - circle(at right 80vw bottom 70%) assert_equals: expected "circle(at calc(100% - 80vw) calc(30%))" but got "circle(at right 80vw bottom 70%)"
FAIL test unit (inline): vw - circle(at right 80vw bottom 70vw) assert_equals: expected "circle(at calc(100% - 80vw) calc(100% - 70vw))" but got "circle(at right 80vw bottom 70vw)"
FAIL test unit (inline): vh - circle(at 50vh) assert_equals: expected "circle(at 50vh 50%)" but got "circle(at 50vh center)"
PASS test unit (inline): vh - circle(at 50% 50vh)
PASS test unit (inline): vh - circle(at 50vh 50%)
PASS test unit (inline): vh - circle(at 50vh 50vh)
FAIL test unit (inline): vh - circle(at left 50vh) assert_equals: expected "circle(at 0% 50vh)" but got "circle(at left 50vh)"
FAIL test unit (inline): vh - circle(at 50vh top) assert_equals: expected "circle(at 50vh 0%)" but got "circle(at 50vh top)"
FAIL test unit (inline): vh - circle(at right 80vh) assert_equals: expected "circle(at 100% 80vh)" but got "circle(at right 80vh)"
FAIL test unit (inline): vh - circle(at 70vh bottom) assert_equals: expected "circle(at 70vh 100%)" but got "circle(at 70vh bottom)"
FAIL test unit (inline): vh - circle(at center 60vh) assert_equals: expected "circle(at 50% 60vh)" but got "circle(at center 60vh)"
FAIL test unit (inline): vh - circle(at 60vh center) assert_equals: expected "circle(at 60vh 50%)" but got "circle(at 60vh center)"
FAIL test unit (inline): vh - circle(at left 50% top 50vh) assert_equals: expected "circle(at 50% 50vh)" but got "circle(at left 50% top 50vh)"
FAIL test unit (inline): vh - circle(at left 50% bottom 70vh) assert_equals: expected "circle(at 50% calc(100% - 70vh))" but got "circle(at left 50% bottom 70vh)"
FAIL test unit (inline): vh - circle(at left 50vh top 50%) assert_equals: expected "circle(at 50vh 50%)" but got "circle(at left 50vh top 50%)"
FAIL test unit (inline): vh - circle(at left 50vh top 50vh) assert_equals: expected "circle(at 50vh 50vh)" but got "circle(at left 50vh top 50vh)"
FAIL test unit (inline): vh - circle(at left 50vh bottom 70%) assert_equals: expected "circle(at 50vh calc(30%))" but got "circle(at left 50vh bottom 70%)"
FAIL test unit (inline): vh - circle(at top 50% left 50vh) assert_equals: expected "circle(at 50vh 50%)" but got "circle(at left 50vh top 50%)"
FAIL test unit (inline): vh - circle(at top 50% right 80vh) assert_equals: expected "circle(at calc(100% - 80vh) 50%)" but got "circle(at right 80vh top 50%)"
FAIL test unit (inline): vh - circle(at top 50vh left 50%) assert_equals: expected "circle(at 50% 50vh)" but got "circle(at left 50% top 50vh)"
FAIL test unit (inline): vh - circle(at top 50vh left 50vh) assert_equals: expected "circle(at 50vh 50vh)" but got "circle(at left 50vh top 50vh)"
FAIL test unit (inline): vh - circle(at top 50vh right 80%) assert_equals: expected "circle(at calc(20%) 50vh)" but got "circle(at right 80% top 50vh)"
FAIL test unit (inline): vh - circle(at bottom 70% left 50vh) assert_equals: expected "circle(at 50vh calc(30%))" but got "circle(at left 50vh bottom 70%)"
FAIL test unit (inline): vh - circle(at bottom 70% right 80vh) assert_equals: expected "circle(at calc(100% - 80vh) calc(30%))" but got "circle(at right 80vh bottom 70%)"
FAIL test unit (inline): vh - circle(at bottom 70vh left 50%) assert_equals: expected "circle(at 50% calc(100% - 70vh))" but got "circle(at left 50% bottom 70vh)"
FAIL test unit (inline): vh - circle(at bottom 70vh right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70vh))" but got "circle(at right 50% bottom 70vh)"
FAIL test unit (inline): vh - circle(at bottom 70vh right 80vh) assert_equals: expected "circle(at calc(100% - 80vh) calc(100% - 70vh))" but got "circle(at right 80vh bottom 70vh)"
FAIL test unit (inline): vh - circle(at right 80% top 50vh) assert_equals: expected "circle(at calc(20%) 50vh)" but got "circle(at right 80% top 50vh)"
FAIL test unit (inline): vh - circle(at right 80% bottom 70vh) assert_equals: expected "circle(at calc(20%) calc(100% - 70vh))" but got "circle(at right 80% bottom 70vh)"
FAIL test unit (inline): vh - circle(at right 80vh top 50%) assert_equals: expected "circle(at calc(100% - 80vh) 50%)" but got "circle(at right 80vh top 50%)"
FAIL test unit (inline): vh - circle(at right 80vh bottom 70%) assert_equals: expected "circle(at calc(100% - 80vh) calc(30%))" but got "circle(at right 80vh bottom 70%)"
FAIL test unit (inline): vh - circle(at right 80vh bottom 70vh) assert_equals: expected "circle(at calc(100% - 80vh) calc(100% - 70vh))" but got "circle(at right 80vh bottom 70vh)"
FAIL test unit (inline): vmin - circle(at 50vmin) assert_equals: expected "circle(at 50vmin 50%)" but got "circle(at 50vmin center)"
PASS test unit (inline): vmin - circle(at 50% 50vmin)
PASS test unit (inline): vmin - circle(at 50vmin 50%)
PASS test unit (inline): vmin - circle(at 50vmin 50vmin)
FAIL test unit (inline): vmin - circle(at left 50vmin) assert_equals: expected "circle(at 0% 50vmin)" but got "circle(at left 50vmin)"
FAIL test unit (inline): vmin - circle(at 50vmin top) assert_equals: expected "circle(at 50vmin 0%)" but got "circle(at 50vmin top)"
FAIL test unit (inline): vmin - circle(at right 80vmin) assert_equals: expected "circle(at 100% 80vmin)" but got "circle(at right 80vmin)"
FAIL test unit (inline): vmin - circle(at 70vmin bottom) assert_equals: expected "circle(at 70vmin 100%)" but got "circle(at 70vmin bottom)"
FAIL test unit (inline): vmin - circle(at center 60vmin) assert_equals: expected "circle(at 50% 60vmin)" but got "circle(at center 60vmin)"
FAIL test unit (inline): vmin - circle(at 60vmin center) assert_equals: expected "circle(at 60vmin 50%)" but got "circle(at 60vmin center)"
FAIL test unit (inline): vmin - circle(at left 50% top 50vmin) assert_equals: expected "circle(at 50% 50vmin)" but got "circle(at left 50% top 50vmin)"
FAIL test unit (inline): vmin - circle(at left 50% bottom 70vmin) assert_equals: expected "circle(at 50% calc(100% - 70vmin))" but got "circle(at left 50% bottom 70vmin)"
FAIL test unit (inline): vmin - circle(at left 50vmin top 50%) assert_equals: expected "circle(at 50vmin 50%)" but got "circle(at left 50vmin top 50%)"
FAIL test unit (inline): vmin - circle(at left 50vmin top 50vmin) assert_equals: expected "circle(at 50vmin 50vmin)" but got "circle(at left 50vmin top 50vmin)"
FAIL test unit (inline): vmin - circle(at left 50vmin bottom 70%) assert_equals: expected "circle(at 50vmin calc(30%))" but got "circle(at left 50vmin bottom 70%)"
FAIL test unit (inline): vmin - circle(at top 50% left 50vmin) assert_equals: expected "circle(at 50vmin 50%)" but got "circle(at left 50vmin top 50%)"
FAIL test unit (inline): vmin - circle(at top 50% right 80vmin) assert_equals: expected "circle(at calc(100% - 80vmin) 50%)" but got "circle(at right 80vmin top 50%)"
FAIL test unit (inline): vmin - circle(at top 50vmin left 50%) assert_equals: expected "circle(at 50% 50vmin)" but got "circle(at left 50% top 50vmin)"
FAIL test unit (inline): vmin - circle(at top 50vmin left 50vmin) assert_equals: expected "circle(at 50vmin 50vmin)" but got "circle(at left 50vmin top 50vmin)"
FAIL test unit (inline): vmin - circle(at top 50vmin right 80%) assert_equals: expected "circle(at calc(20%) 50vmin)" but got "circle(at right 80% top 50vmin)"
FAIL test unit (inline): vmin - circle(at bottom 70% left 50vmin) assert_equals: expected "circle(at 50vmin calc(30%))" but got "circle(at left 50vmin bottom 70%)"
FAIL test unit (inline): vmin - circle(at bottom 70% right 80vmin) assert_equals: expected "circle(at calc(100% - 80vmin) calc(30%))" but got "circle(at right 80vmin bottom 70%)"
FAIL test unit (inline): vmin - circle(at bottom 70vmin left 50%) assert_equals: expected "circle(at 50% calc(100% - 70vmin))" but got "circle(at left 50% bottom 70vmin)"
FAIL test unit (inline): vmin - circle(at bottom 70vmin right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70vmin))" but got "circle(at right 50% bottom 70vmin)"
FAIL test unit (inline): vmin - circle(at bottom 70vmin right 80vmin) assert_equals: expected "circle(at calc(100% - 80vmin) calc(100% - 70vmin))" but got "circle(at right 80vmin bottom 70vmin)"
FAIL test unit (inline): vmin - circle(at right 80% top 50vmin) assert_equals: expected "circle(at calc(20%) 50vmin)" but got "circle(at right 80% top 50vmin)"
FAIL test unit (inline): vmin - circle(at right 80% bottom 70vmin) assert_equals: expected "circle(at calc(20%) calc(100% - 70vmin))" but got "circle(at right 80% bottom 70vmin)"
FAIL test unit (inline): vmin - circle(at right 80vmin top 50%) assert_equals: expected "circle(at calc(100% - 80vmin) 50%)" but got "circle(at right 80vmin top 50%)"
FAIL test unit (inline): vmin - circle(at right 80vmin bottom 70%) assert_equals: expected "circle(at calc(100% - 80vmin) calc(30%))" but got "circle(at right 80vmin bottom 70%)"
FAIL test unit (inline): vmin - circle(at right 80vmin bottom 70vmin) assert_equals: expected "circle(at calc(100% - 80vmin) calc(100% - 70vmin))" but got "circle(at right 80vmin bottom 70vmin)"
FAIL test unit (inline): vmax - circle(at 50vmax) assert_equals: expected "circle(at 50vmax 50%)" but got "circle(at 50vmax center)"
PASS test unit (inline): vmax - circle(at 50% 50vmax)
PASS test unit (inline): vmax - circle(at 50vmax 50%)
PASS test unit (inline): vmax - circle(at 50vmax 50vmax)
FAIL test unit (inline): vmax - circle(at left 50vmax) assert_equals: expected "circle(at 0% 50vmax)" but got "circle(at left 50vmax)"
FAIL test unit (inline): vmax - circle(at 50vmax top) assert_equals: expected "circle(at 50vmax 0%)" but got "circle(at 50vmax top)"
FAIL test unit (inline): vmax - circle(at right 80vmax) assert_equals: expected "circle(at 100% 80vmax)" but got "circle(at right 80vmax)"
FAIL test unit (inline): vmax - circle(at 70vmax bottom) assert_equals: expected "circle(at 70vmax 100%)" but got "circle(at 70vmax bottom)"
FAIL test unit (inline): vmax - circle(at center 60vmax) assert_equals: expected "circle(at 50% 60vmax)" but got "circle(at center 60vmax)"
FAIL test unit (inline): vmax - circle(at 60vmax center) assert_equals: expected "circle(at 60vmax 50%)" but got "circle(at 60vmax center)"
FAIL test unit (inline): vmax - circle(at left 50% top 50vmax) assert_equals: expected "circle(at 50% 50vmax)" but got "circle(at left 50% top 50vmax)"
FAIL test unit (inline): vmax - circle(at left 50% bottom 70vmax) assert_equals: expected "circle(at 50% calc(100% - 70vmax))" but got "circle(at left 50% bottom 70vmax)"
FAIL test unit (inline): vmax - circle(at left 50vmax top 50%) assert_equals: expected "circle(at 50vmax 50%)" but got "circle(at left 50vmax top 50%)"
FAIL test unit (inline): vmax - circle(at left 50vmax top 50vmax) assert_equals: expected "circle(at 50vmax 50vmax)" but got "circle(at left 50vmax top 50vmax)"
FAIL test unit (inline): vmax - circle(at left 50vmax bottom 70%) assert_equals: expected "circle(at 50vmax calc(30%))" but got "circle(at left 50vmax bottom 70%)"
FAIL test unit (inline): vmax - circle(at top 50% left 50vmax) assert_equals: expected "circle(at 50vmax 50%)" but got "circle(at left 50vmax top 50%)"
FAIL test unit (inline): vmax - circle(at top 50% right 80vmax) assert_equals: expected "circle(at calc(100% - 80vmax) 50%)" but got "circle(at right 80vmax top 50%)"
FAIL test unit (inline): vmax - circle(at top 50vmax left 50%) assert_equals: expected "circle(at 50% 50vmax)" but got "circle(at left 50% top 50vmax)"
FAIL test unit (inline): vmax - circle(at top 50vmax left 50vmax) assert_equals: expected "circle(at 50vmax 50vmax)" but got "circle(at left 50vmax top 50vmax)"
FAIL test unit (inline): vmax - circle(at top 50vmax right 80%) assert_equals: expected "circle(at calc(20%) 50vmax)" but got "circle(at right 80% top 50vmax)"
FAIL test unit (inline): vmax - circle(at bottom 70% left 50vmax) assert_equals: expected "circle(at 50vmax calc(30%))" but got "circle(at left 50vmax bottom 70%)"
FAIL test unit (inline): vmax - circle(at bottom 70% right 80vmax) assert_equals: expected "circle(at calc(100% - 80vmax) calc(30%))" but got "circle(at right 80vmax bottom 70%)"
FAIL test unit (inline): vmax - circle(at bottom 70vmax left 50%) assert_equals: expected "circle(at 50% calc(100% - 70vmax))" but got "circle(at left 50% bottom 70vmax)"
FAIL test unit (inline): vmax - circle(at bottom 70vmax right 50%) assert_equals: expected "circle(at calc(50%) calc(100% - 70vmax))" but got "circle(at right 50% bottom 70vmax)"
FAIL test unit (inline): vmax - circle(at bottom 70vmax right 80vmax) assert_equals: expected "circle(at calc(100% - 80vmax) calc(100% - 70vmax))" but got "circle(at right 80vmax bottom 70vmax)"
FAIL test unit (inline): vmax - circle(at right 80% top 50vmax) assert_equals: expected "circle(at calc(20%) 50vmax)" but got "circle(at right 80% top 50vmax)"
FAIL test unit (inline): vmax - circle(at right 80% bottom 70vmax) assert_equals: expected "circle(at calc(20%) calc(100% - 70vmax))" but got "circle(at right 80% bottom 70vmax)"
FAIL test unit (inline): vmax - circle(at right 80vmax top 50%) assert_equals: expected "circle(at calc(100% - 80vmax) 50%)" but got "circle(at right 80vmax top 50%)"
FAIL test unit (inline): vmax - circle(at right 80vmax bottom 70%) assert_equals: expected "circle(at calc(100% - 80vmax) calc(30%))" but got "circle(at right 80vmax bottom 70%)"
FAIL test unit (inline): vmax - circle(at right 80vmax bottom 70vmax) assert_equals: expected "circle(at calc(100% - 80vmax) calc(100% - 70vmax))" but got "circle(at right 80vmax bottom 70vmax)"
PASS test unit (computed): cm - circle(at 50cm)
PASS test unit (computed): cm - circle(at 50% 50cm)
PASS test unit (computed): cm - circle(at 50cm 50%)
PASS test unit (computed): cm - circle(at 50cm 50cm)
PASS test unit (computed): cm - circle(at left 50cm)
PASS test unit (computed): cm - circle(at 50cm top)
PASS test unit (computed): cm - circle(at right 80cm)
PASS test unit (computed): cm - circle(at 70cm bottom)
PASS test unit (computed): cm - circle(at center 60cm)
PASS test unit (computed): cm - circle(at 60cm center)
PASS test unit (computed): cm - circle(at left 50% top 50cm)
PASS test unit (computed): cm - circle(at left 50% bottom 70cm)
PASS test unit (computed): cm - circle(at left 50cm top 50%)
PASS test unit (computed): cm - circle(at left 50cm top 50cm)
PASS test unit (computed): cm - circle(at left 50cm bottom 70%)
PASS test unit (computed): cm - circle(at top 50% left 50cm)
PASS test unit (computed): cm - circle(at top 50% right 80cm)
PASS test unit (computed): cm - circle(at top 50cm left 50%)
PASS test unit (computed): cm - circle(at top 50cm left 50cm)
PASS test unit (computed): cm - circle(at top 50cm right 80%)
PASS test unit (computed): cm - circle(at bottom 70% left 50cm)
PASS test unit (computed): cm - circle(at bottom 70% right 80cm)
PASS test unit (computed): cm - circle(at bottom 70cm left 50%)
PASS test unit (computed): cm - circle(at bottom 70cm right 50%)
PASS test unit (computed): cm - circle(at bottom 70cm right 80cm)
PASS test unit (computed): cm - circle(at right 80% top 50cm)
PASS test unit (computed): cm - circle(at right 80% bottom 70cm)
PASS test unit (computed): cm - circle(at right 80cm top 50%)
PASS test unit (computed): cm - circle(at right 80cm bottom 70%)
PASS test unit (computed): cm - circle(at right 80cm bottom 70cm)
PASS test unit (computed): mm - circle(at 50mm)
PASS test unit (computed): mm - circle(at 50% 50mm)
PASS test unit (computed): mm - circle(at 50mm 50%)
PASS test unit (computed): mm - circle(at 50mm 50mm)
PASS test unit (computed): mm - circle(at left 50mm)
PASS test unit (computed): mm - circle(at 50mm top)
PASS test unit (computed): mm - circle(at right 80mm)
PASS test unit (computed): mm - circle(at 70mm bottom)
PASS test unit (computed): mm - circle(at center 60mm)
PASS test unit (computed): mm - circle(at 60mm center)
PASS test unit (computed): mm - circle(at left 50% top 50mm)
PASS test unit (computed): mm - circle(at left 50% bottom 70mm)
PASS test unit (computed): mm - circle(at left 50mm top 50%)
PASS test unit (computed): mm - circle(at left 50mm top 50mm)
PASS test unit (computed): mm - circle(at left 50mm bottom 70%)
PASS test unit (computed): mm - circle(at top 50% left 50mm)
PASS test unit (computed): mm - circle(at top 50% right 80mm)
PASS test unit (computed): mm - circle(at top 50mm left 50%)
PASS test unit (computed): mm - circle(at top 50mm left 50mm)
PASS test unit (computed): mm - circle(at top 50mm right 80%)
PASS test unit (computed): mm - circle(at bottom 70% left 50mm)
PASS test unit (computed): mm - circle(at bottom 70% right 80mm)
PASS test unit (computed): mm - circle(at bottom 70mm left 50%)
PASS test unit (computed): mm - circle(at bottom 70mm right 50%)
PASS test unit (computed): mm - circle(at bottom 70mm right 80mm)
PASS test unit (computed): mm - circle(at right 80% top 50mm)
PASS test unit (computed): mm - circle(at right 80% bottom 70mm)
PASS test unit (computed): mm - circle(at right 80mm top 50%)
PASS test unit (computed): mm - circle(at right 80mm bottom 70%)
PASS test unit (computed): mm - circle(at right 80mm bottom 70mm)
PASS test unit (computed): in - circle(at 50in)
PASS test unit (computed): in - circle(at 50% 50in)
PASS test unit (computed): in - circle(at 50in 50%)
PASS test unit (computed): in - circle(at 50in 50in)
PASS test unit (computed): in - circle(at left 50in)
PASS test unit (computed): in - circle(at 50in top)
PASS test unit (computed): in - circle(at right 80in)
PASS test unit (computed): in - circle(at 70in bottom)
PASS test unit (computed): in - circle(at center 60in)
PASS test unit (computed): in - circle(at 60in center)
PASS test unit (computed): in - circle(at left 50% top 50in)
PASS test unit (computed): in - circle(at left 50% bottom 70in)
PASS test unit (computed): in - circle(at left 50in top 50%)
PASS test unit (computed): in - circle(at left 50in top 50in)
PASS test unit (computed): in - circle(at left 50in bottom 70%)
PASS test unit (computed): in - circle(at top 50% left 50in)
PASS test unit (computed): in - circle(at top 50% right 80in)
PASS test unit (computed): in - circle(at top 50in left 50%)
PASS test unit (computed): in - circle(at top 50in left 50in)
PASS test unit (computed): in - circle(at top 50in right 80%)
PASS test unit (computed): in - circle(at bottom 70% left 50in)
PASS test unit (computed): in - circle(at bottom 70% right 80in)
PASS test unit (computed): in - circle(at bottom 70in left 50%)
PASS test unit (computed): in - circle(at bottom 70in right 50%)
PASS test unit (computed): in - circle(at bottom 70in right 80in)
PASS test unit (computed): in - circle(at right 80% top 50in)
PASS test unit (computed): in - circle(at right 80% bottom 70in)
PASS test unit (computed): in - circle(at right 80in top 50%)
PASS test unit (computed): in - circle(at right 80in bottom 70%)
PASS test unit (computed): in - circle(at right 80in bottom 70in)
PASS test unit (computed): pt - circle(at 50pt)
PASS test unit (computed): pt - circle(at 50% 50pt)
PASS test unit (computed): pt - circle(at 50pt 50%)
PASS test unit (computed): pt - circle(at 50pt 50pt)
PASS test unit (computed): pt - circle(at left 50pt)
PASS test unit (computed): pt - circle(at 50pt top)
PASS test unit (computed): pt - circle(at right 80pt)
PASS test unit (computed): pt - circle(at 70pt bottom)
PASS test unit (computed): pt - circle(at center 60pt)
PASS test unit (computed): pt - circle(at 60pt center)
PASS test unit (computed): pt - circle(at left 50% top 50pt)
PASS test unit (computed): pt - circle(at left 50% bottom 70pt)
PASS test unit (computed): pt - circle(at left 50pt top 50%)
PASS test unit (computed): pt - circle(at left 50pt top 50pt)
PASS test unit (computed): pt - circle(at left 50pt bottom 70%)
PASS test unit (computed): pt - circle(at top 50% left 50pt)
PASS test unit (computed): pt - circle(at top 50% right 80pt)
PASS test unit (computed): pt - circle(at top 50pt left 50%)
PASS test unit (computed): pt - circle(at top 50pt left 50pt)
PASS test unit (computed): pt - circle(at top 50pt right 80%)
PASS test unit (computed): pt - circle(at bottom 70% left 50pt)
PASS test unit (computed): pt - circle(at bottom 70% right 80pt)
PASS test unit (computed): pt - circle(at bottom 70pt left 50%)
PASS test unit (computed): pt - circle(at bottom 70pt right 50%)
PASS test unit (computed): pt - circle(at bottom 70pt right 80pt)
PASS test unit (computed): pt - circle(at right 80% top 50pt)
PASS test unit (computed): pt - circle(at right 80% bottom 70pt)
PASS test unit (computed): pt - circle(at right 80pt top 50%)
PASS test unit (computed): pt - circle(at right 80pt bottom 70%)
PASS test unit (computed): pt - circle(at right 80pt bottom 70pt)
PASS test unit (computed): pc - circle(at 50pc)
PASS test unit (computed): pc - circle(at 50% 50pc)
PASS test unit (computed): pc - circle(at 50pc 50%)
PASS test unit (computed): pc - circle(at 50pc 50pc)
PASS test unit (computed): pc - circle(at left 50pc)
PASS test unit (computed): pc - circle(at 50pc top)
PASS test unit (computed): pc - circle(at right 80pc)
PASS test unit (computed): pc - circle(at 70pc bottom)
PASS test unit (computed): pc - circle(at center 60pc)
PASS test unit (computed): pc - circle(at 60pc center)
PASS test unit (computed): pc - circle(at left 50% top 50pc)
PASS test unit (computed): pc - circle(at left 50% bottom 70pc)
PASS test unit (computed): pc - circle(at left 50pc top 50%)
PASS test unit (computed): pc - circle(at left 50pc top 50pc)
PASS test unit (computed): pc - circle(at left 50pc bottom 70%)
PASS test unit (computed): pc - circle(at top 50% left 50pc)
PASS test unit (computed): pc - circle(at top 50% right 80pc)
PASS test unit (computed): pc - circle(at top 50pc left 50%)
PASS test unit (computed): pc - circle(at top 50pc left 50pc)
PASS test unit (computed): pc - circle(at top 50pc right 80%)
PASS test unit (computed): pc - circle(at bottom 70% left 50pc)
PASS test unit (computed): pc - circle(at bottom 70% right 80pc)
PASS test unit (computed): pc - circle(at bottom 70pc left 50%)
PASS test unit (computed): pc - circle(at bottom 70pc right 50%)
PASS test unit (computed): pc - circle(at bottom 70pc right 80pc)
PASS test unit (computed): pc - circle(at right 80% top 50pc)
PASS test unit (computed): pc - circle(at right 80% bottom 70pc)
PASS test unit (computed): pc - circle(at right 80pc top 50%)
PASS test unit (computed): pc - circle(at right 80pc bottom 70%)
PASS test unit (computed): pc - circle(at right 80pc bottom 70pc)
PASS test unit (computed): em - circle(at 50em)
PASS test unit (computed): em - circle(at 50% 50em)
PASS test unit (computed): em - circle(at 50em 50%)
PASS test unit (computed): em - circle(at 50em 50em)
PASS test unit (computed): em - circle(at left 50em)
PASS test unit (computed): em - circle(at 50em top)
PASS test unit (computed): em - circle(at right 80em)
PASS test unit (computed): em - circle(at 70em bottom)
PASS test unit (computed): em - circle(at center 60em)
PASS test unit (computed): em - circle(at 60em center)
PASS test unit (computed): em - circle(at left 50% top 50em)
PASS test unit (computed): em - circle(at left 50% bottom 70em)
PASS test unit (computed): em - circle(at left 50em top 50%)
PASS test unit (computed): em - circle(at left 50em top 50em)
PASS test unit (computed): em - circle(at left 50em bottom 70%)
PASS test unit (computed): em - circle(at top 50% left 50em)
PASS test unit (computed): em - circle(at top 50% right 80em)
PASS test unit (computed): em - circle(at top 50em left 50%)
PASS test unit (computed): em - circle(at top 50em left 50em)
PASS test unit (computed): em - circle(at top 50em right 80%)
PASS test unit (computed): em - circle(at bottom 70% left 50em)
PASS test unit (computed): em - circle(at bottom 70% right 80em)
PASS test unit (computed): em - circle(at bottom 70em left 50%)
PASS test unit (computed): em - circle(at bottom 70em right 50%)
PASS test unit (computed): em - circle(at bottom 70em right 80em)
PASS test unit (computed): em - circle(at right 80% top 50em)
PASS test unit (computed): em - circle(at right 80% bottom 70em)
PASS test unit (computed): em - circle(at right 80em top 50%)
PASS test unit (computed): em - circle(at right 80em bottom 70%)
PASS test unit (computed): em - circle(at right 80em bottom 70em)
PASS test unit (computed): ex - circle(at 50ex)
PASS test unit (computed): ex - circle(at 50% 50ex)
PASS test unit (computed): ex - circle(at 50ex 50%)
PASS test unit (computed): ex - circle(at 50ex 50ex)
PASS test unit (computed): ex - circle(at left 50ex)
PASS test unit (computed): ex - circle(at 50ex top)
PASS test unit (computed): ex - circle(at right 80ex)
PASS test unit (computed): ex - circle(at 70ex bottom)
PASS test unit (computed): ex - circle(at center 60ex)
PASS test unit (computed): ex - circle(at 60ex center)
PASS test unit (computed): ex - circle(at left 50% top 50ex)
PASS test unit (computed): ex - circle(at left 50% bottom 70ex)
PASS test unit (computed): ex - circle(at left 50ex top 50%)
PASS test unit (computed): ex - circle(at left 50ex top 50ex)
PASS test unit (computed): ex - circle(at left 50ex bottom 70%)
PASS test unit (computed): ex - circle(at top 50% left 50ex)
PASS test unit (computed): ex - circle(at top 50% right 80ex)
PASS test unit (computed): ex - circle(at top 50ex left 50%)
PASS test unit (computed): ex - circle(at top 50ex left 50ex)
PASS test unit (computed): ex - circle(at top 50ex right 80%)
PASS test unit (computed): ex - circle(at bottom 70% left 50ex)
PASS test unit (computed): ex - circle(at bottom 70% right 80ex)
PASS test unit (computed): ex - circle(at bottom 70ex left 50%)
PASS test unit (computed): ex - circle(at bottom 70ex right 50%)
PASS test unit (computed): ex - circle(at bottom 70ex right 80ex)
PASS test unit (computed): ex - circle(at right 80% top 50ex)
PASS test unit (computed): ex - circle(at right 80% bottom 70ex)
PASS test unit (computed): ex - circle(at right 80ex top 50%)
PASS test unit (computed): ex - circle(at right 80ex bottom 70%)
PASS test unit (computed): ex - circle(at right 80ex bottom 70ex)
PASS test unit (computed): ch - circle(at 50ch)
PASS test unit (computed): ch - circle(at 50% 50ch)
PASS test unit (computed): ch - circle(at 50ch 50%)
PASS test unit (computed): ch - circle(at 50ch 50ch)
PASS test unit (computed): ch - circle(at left 50ch)
PASS test unit (computed): ch - circle(at 50ch top)
PASS test unit (computed): ch - circle(at right 80ch)
PASS test unit (computed): ch - circle(at 70ch bottom)
PASS test unit (computed): ch - circle(at center 60ch)
PASS test unit (computed): ch - circle(at 60ch center)
PASS test unit (computed): ch - circle(at left 50% top 50ch)
PASS test unit (computed): ch - circle(at left 50% bottom 70ch)
PASS test unit (computed): ch - circle(at left 50ch top 50%)
PASS test unit (computed): ch - circle(at left 50ch top 50ch)
PASS test unit (computed): ch - circle(at left 50ch bottom 70%)
PASS test unit (computed): ch - circle(at top 50% left 50ch)
PASS test unit (computed): ch - circle(at top 50% right 80ch)
PASS test unit (computed): ch - circle(at top 50ch left 50%)
PASS test unit (computed): ch - circle(at top 50ch left 50ch)
PASS test unit (computed): ch - circle(at top 50ch right 80%)
PASS test unit (computed): ch - circle(at bottom 70% left 50ch)
PASS test unit (computed): ch - circle(at bottom 70% right 80ch)
PASS test unit (computed): ch - circle(at bottom 70ch left 50%)
PASS test unit (computed): ch - circle(at bottom 70ch right 50%)
PASS test unit (computed): ch - circle(at bottom 70ch right 80ch)
PASS test unit (computed): ch - circle(at right 80% top 50ch)
PASS test unit (computed): ch - circle(at right 80% bottom 70ch)
PASS test unit (computed): ch - circle(at right 80ch top 50%)
PASS test unit (computed): ch - circle(at right 80ch bottom 70%)
PASS test unit (computed): ch - circle(at right 80ch bottom 70ch)
PASS test unit (computed): rem - circle(at 50rem)
PASS test unit (computed): rem - circle(at 50% 50rem)
PASS test unit (computed): rem - circle(at 50rem 50%)
PASS test unit (computed): rem - circle(at 50rem 50rem)
PASS test unit (computed): rem - circle(at left 50rem)
PASS test unit (computed): rem - circle(at 50rem top)
PASS test unit (computed): rem - circle(at right 80rem)
PASS test unit (computed): rem - circle(at 70rem bottom)
PASS test unit (computed): rem - circle(at center 60rem)
PASS test unit (computed): rem - circle(at 60rem center)
PASS test unit (computed): rem - circle(at left 50% top 50rem)
PASS test unit (computed): rem - circle(at left 50% bottom 70rem)
PASS test unit (computed): rem - circle(at left 50rem top 50%)
PASS test unit (computed): rem - circle(at left 50rem top 50rem)
PASS test unit (computed): rem - circle(at left 50rem bottom 70%)
PASS test unit (computed): rem - circle(at top 50% left 50rem)
PASS test unit (computed): rem - circle(at top 50% right 80rem)
PASS test unit (computed): rem - circle(at top 50rem left 50%)
PASS test unit (computed): rem - circle(at top 50rem left 50rem)
PASS test unit (computed): rem - circle(at top 50rem right 80%)
PASS test unit (computed): rem - circle(at bottom 70% left 50rem)
PASS test unit (computed): rem - circle(at bottom 70% right 80rem)
PASS test unit (computed): rem - circle(at bottom 70rem left 50%)
PASS test unit (computed): rem - circle(at bottom 70rem right 50%)
PASS test unit (computed): rem - circle(at bottom 70rem right 80rem)
PASS test unit (computed): rem - circle(at right 80% top 50rem)
PASS test unit (computed): rem - circle(at right 80% bottom 70rem)
PASS test unit (computed): rem - circle(at right 80rem top 50%)
PASS test unit (computed): rem - circle(at right 80rem bottom 70%)
PASS test unit (computed): rem - circle(at right 80rem bottom 70rem)
PASS test unit (computed): vw - circle(at 50vw)
PASS test unit (computed): vw - circle(at 50% 50vw)
PASS test unit (computed): vw - circle(at 50vw 50%)
PASS test unit (computed): vw - circle(at 50vw 50vw)
PASS test unit (computed): vw - circle(at left 50vw)
PASS test unit (computed): vw - circle(at 50vw top)
PASS test unit (computed): vw - circle(at right 80vw)
PASS test unit (computed): vw - circle(at 70vw bottom)
PASS test unit (computed): vw - circle(at center 60vw)
PASS test unit (computed): vw - circle(at 60vw center)
PASS test unit (computed): vw - circle(at left 50% top 50vw)
PASS test unit (computed): vw - circle(at left 50% bottom 70vw)
PASS test unit (computed): vw - circle(at left 50vw top 50%)
PASS test unit (computed): vw - circle(at left 50vw top 50vw)
PASS test unit (computed): vw - circle(at left 50vw bottom 70%)
PASS test unit (computed): vw - circle(at top 50% left 50vw)
PASS test unit (computed): vw - circle(at top 50% right 80vw)
PASS test unit (computed): vw - circle(at top 50vw left 50%)
PASS test unit (computed): vw - circle(at top 50vw left 50vw)
PASS test unit (computed): vw - circle(at top 50vw right 80%)
PASS test unit (computed): vw - circle(at bottom 70% left 50vw)
PASS test unit (computed): vw - circle(at bottom 70% right 80vw)
PASS test unit (computed): vw - circle(at bottom 70vw left 50%)
PASS test unit (computed): vw - circle(at bottom 70vw right 50%)
PASS test unit (computed): vw - circle(at bottom 70vw right 80vw)
PASS test unit (computed): vw - circle(at right 80% top 50vw)
PASS test unit (computed): vw - circle(at right 80% bottom 70vw)
PASS test unit (computed): vw - circle(at right 80vw top 50%)
PASS test unit (computed): vw - circle(at right 80vw bottom 70%)
PASS test unit (computed): vw - circle(at right 80vw bottom 70vw)
PASS test unit (computed): vh - circle(at 50vh)
PASS test unit (computed): vh - circle(at 50% 50vh)
PASS test unit (computed): vh - circle(at 50vh 50%)
PASS test unit (computed): vh - circle(at 50vh 50vh)
PASS test unit (computed): vh - circle(at left 50vh)
PASS test unit (computed): vh - circle(at 50vh top)
PASS test unit (computed): vh - circle(at right 80vh)
PASS test unit (computed): vh - circle(at 70vh bottom)
PASS test unit (computed): vh - circle(at center 60vh)
PASS test unit (computed): vh - circle(at 60vh center)
PASS test unit (computed): vh - circle(at left 50% top 50vh)
PASS test unit (computed): vh - circle(at left 50% bottom 70vh)
PASS test unit (computed): vh - circle(at left 50vh top 50%)
PASS test unit (computed): vh - circle(at left 50vh top 50vh)
PASS test unit (computed): vh - circle(at left 50vh bottom 70%)
PASS test unit (computed): vh - circle(at top 50% left 50vh)
PASS test unit (computed): vh - circle(at top 50% right 80vh)
PASS test unit (computed): vh - circle(at top 50vh left 50%)
PASS test unit (computed): vh - circle(at top 50vh left 50vh)
PASS test unit (computed): vh - circle(at top 50vh right 80%)
PASS test unit (computed): vh - circle(at bottom 70% left 50vh)
PASS test unit (computed): vh - circle(at bottom 70% right 80vh)
PASS test unit (computed): vh - circle(at bottom 70vh left 50%)
PASS test unit (computed): vh - circle(at bottom 70vh right 50%)
PASS test unit (computed): vh - circle(at bottom 70vh right 80vh)
PASS test unit (computed): vh - circle(at right 80% top 50vh)
PASS test unit (computed): vh - circle(at right 80% bottom 70vh)
PASS test unit (computed): vh - circle(at right 80vh top 50%)
PASS test unit (computed): vh - circle(at right 80vh bottom 70%)
PASS test unit (computed): vh - circle(at right 80vh bottom 70vh)
PASS test unit (computed): vmin - circle(at 50vmin)
PASS test unit (computed): vmin - circle(at 50% 50vmin)
PASS test unit (computed): vmin - circle(at 50vmin 50%)
PASS test unit (computed): vmin - circle(at 50vmin 50vmin)
PASS test unit (computed): vmin - circle(at left 50vmin)
PASS test unit (computed): vmin - circle(at 50vmin top)
PASS test unit (computed): vmin - circle(at right 80vmin)
PASS test unit (computed): vmin - circle(at 70vmin bottom)
PASS test unit (computed): vmin - circle(at center 60vmin)
PASS test unit (computed): vmin - circle(at 60vmin center)
PASS test unit (computed): vmin - circle(at left 50% top 50vmin)
PASS test unit (computed): vmin - circle(at left 50% bottom 70vmin)
PASS test unit (computed): vmin - circle(at left 50vmin top 50%)
PASS test unit (computed): vmin - circle(at left 50vmin top 50vmin)
PASS test unit (computed): vmin - circle(at left 50vmin bottom 70%)
PASS test unit (computed): vmin - circle(at top 50% left 50vmin)
PASS test unit (computed): vmin - circle(at top 50% right 80vmin)
PASS test unit (computed): vmin - circle(at top 50vmin left 50%)
PASS test unit (computed): vmin - circle(at top 50vmin left 50vmin)
PASS test unit (computed): vmin - circle(at top 50vmin right 80%)
PASS test unit (computed): vmin - circle(at bottom 70% left 50vmin)
PASS test unit (computed): vmin - circle(at bottom 70% right 80vmin)
PASS test unit (computed): vmin - circle(at bottom 70vmin left 50%)
PASS test unit (computed): vmin - circle(at bottom 70vmin right 50%)
PASS test unit (computed): vmin - circle(at bottom 70vmin right 80vmin)
PASS test unit (computed): vmin - circle(at right 80% top 50vmin)
PASS test unit (computed): vmin - circle(at right 80% bottom 70vmin)
PASS test unit (computed): vmin - circle(at right 80vmin top 50%)
PASS test unit (computed): vmin - circle(at right 80vmin bottom 70%)
PASS test unit (computed): vmin - circle(at right 80vmin bottom 70vmin)
PASS test unit (computed): vmax - circle(at 50vmax)
PASS test unit (computed): vmax - circle(at 50% 50vmax)
PASS test unit (computed): vmax - circle(at 50vmax 50%)
PASS test unit (computed): vmax - circle(at 50vmax 50vmax)
PASS test unit (computed): vmax - circle(at left 50vmax)
PASS test unit (computed): vmax - circle(at 50vmax top)
PASS test unit (computed): vmax - circle(at right 80vmax)
PASS test unit (computed): vmax - circle(at 70vmax bottom)
PASS test unit (computed): vmax - circle(at center 60vmax)
PASS test unit (computed): vmax - circle(at 60vmax center)
PASS test unit (computed): vmax - circle(at left 50% top 50vmax)
PASS test unit (computed): vmax - circle(at left 50% bottom 70vmax)
PASS test unit (computed): vmax - circle(at left 50vmax top 50%)
PASS test unit (computed): vmax - circle(at left 50vmax top 50vmax)
PASS test unit (computed): vmax - circle(at left 50vmax bottom 70%)
PASS test unit (computed): vmax - circle(at top 50% left 50vmax)
PASS test unit (computed): vmax - circle(at top 50% right 80vmax)
PASS test unit (computed): vmax - circle(at top 50vmax left 50%)
PASS test unit (computed): vmax - circle(at top 50vmax left 50vmax)
PASS test unit (computed): vmax - circle(at top 50vmax right 80%)
PASS test unit (computed): vmax - circle(at bottom 70% left 50vmax)
PASS test unit (computed): vmax - circle(at bottom 70% right 80vmax)
PASS test unit (computed): vmax - circle(at bottom 70vmax left 50%)
PASS test unit (computed): vmax - circle(at bottom 70vmax right 50%)
PASS test unit (computed): vmax - circle(at bottom 70vmax right 80vmax)
PASS test unit (computed): vmax - circle(at right 80% top 50vmax)
PASS test unit (computed): vmax - circle(at right 80% bottom 70vmax)
PASS test unit (computed): vmax - circle(at right 80vmax top 50%)
PASS test unit (computed): vmax - circle(at right 80vmax bottom 70%)
PASS test unit (computed): vmax - circle(at right 80vmax bottom 70vmax)
